requirements.txt
llama-index==0.9.48
pypdf==3.17.1
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.109.1
uvicorn==0.27.1
pydantic==2.6.1 